+++ src/macros.rs
@@ -416,13 +416,26 @@ macro_rules! assert_snapshot {
 
 /// Settings configuration macro.
 ///
-/// This macro lets you bind some settings temporarily.  The first argument
+/// This macro lets you bind some [`Settings`](crate::Settings) temporarily.  The first argument
 /// takes key value pairs that should be set, the second is the block to
-/// execute.  All settings can be set (`sort_maps => value` maps roughly
-/// to `set_sort_maps(value)`).
+/// execute.  All settings can be set (`sort_maps => value` maps to `set_sort_maps(value)`).
+/// The exception are redactions which can only be set to a vector this way.
+///
+/// This example:
 ///
 /// ```rust
 /// insta::with_settings!({sort_maps => true}, {
+///     // run snapshot test here
+/// });
+/// ```
+///
+/// Is equivalent to the following:
+///
+/// ```rust
+/// # use insta::Settings;
+/// let mut settings = Settings::new();
+/// settings.set_sort_maps(true);
+/// settings.bind(|| {
 ///     // run snapshot test here
 /// });
 /// ```
@@ -431,7 +444,7 @@ macro_rules! with_settings {
     ({$($k:ident => $v:expr),*$(,)?}, $body:block) => {{
         let mut settings = $crate::Settings::new();
         $(
-            settings._private_inner_mut().$k = $v.into();
+            settings._private_inner_mut().$k($v);
         )*
         settings.bind(|| $body)
     }}