+++ src/output.rs
@@ -2,7 +2,7 @@ use std::{path::Path, time::Duration};
 
 use similar::{Algorithm, ChangeTag, TextDiff};
 
-use crate::snapshot::Snapshot;
+use crate::snapshot::{MetaData, Snapshot};
 use crate::utils::{format_rust_expression, style, term_width};
 
 /// Prints the summary of a snapshot
@@ -59,6 +59,7 @@ pub fn print_snapshot_diff(
     old_snapshot: Option<&Snapshot>,
     snapshot_file: Option<&Path>,
     mut line: Option<u32>,
+    show_info: bool,
 ) {
     // default to old assertion line from snapshot.
     if line.is_none() {
@@ -68,17 +69,7 @@ pub fn print_snapshot_diff(
     print_snapshot_summary(workspace_root, new, snapshot_file, line);
     let old_contents = old_snapshot.as_ref().map_or("", |x| x.contents_str());
     let new_contents = new.contents_str();
-    if !old_contents.is_empty() {
-        println!("{}", style("-old snapshot").red());
-        println!("{}", style("+new results").green());
-    } else {
-        println!("{}", style("+new results").green());
-    }
-    print_changeset(
-        old_contents,
-        new_contents,
-        new.metadata().expression.as_deref(),
-    );
+    print_changeset(old_contents, new_contents, new.metadata(), show_info);
 }
 
 pub fn print_snapshot_diff_with_title(
@@ -100,6 +91,7 @@ pub fn print_snapshot_diff_with_title(
         old_snapshot,
         snapshot_file,
         Some(line),
+        true,
     );
 }
 
@@ -121,17 +113,39 @@ pub fn print_snapshot_summary_with_title(
     println!("{title:━^width$}", title = "", width = width);
 }
 
-pub fn print_changeset(old: &str, new: &str, expr: Option<&str>) {
+pub fn print_changeset(old: &str, new: &str, metadata: &MetaData, show_info: bool) {
     let width = term_width();
     let diff = TextDiff::configure()
         .algorithm(Algorithm::Patience)
         .timeout(Duration::from_millis(500))
         .diff_lines(old, new);
+    println!("{:─^1$}", "", width);
 
-    if let Some(expr) = expr {
-        println!("{:─^1$}", "", width,);
-        println!("{}", style(format_rust_expression(expr)));
+    if show_info {
+        if let Some(expr) = metadata.expression() {
+            println!("Expression: {}", style(format_rust_expression(expr)));
+            println!("{:─^1$}", "", width);
+        }
+
+        if let Some(descr) = metadata.description() {
+            println!("{}", descr);
+            println!("{:─^1$}", "", width);
+        }
+
+        if let Some(info) = metadata.private_info() {
+            let out = serde_yaml::to_string(&info).unwrap();
+            println!("{}", out.trim().strip_prefix("---").unwrap().trim_start());
+            println!("{:─^1$}", "", width);
+        }
     }
+
+    if !old.is_empty() {
+        println!("{}", style("-old snapshot").red());
+        println!("{}", style("+new results").green());
+    } else {
+        println!("{}", style("+new results").green());
+    }
+
     println!("────────────┬{:─^1$}", "", width.saturating_sub(13));
     let mut has_changes = false;
     for (idx, group) in diff.grouped_ops(4).iter().enumerate() {