+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to insta and cargo-insta are documented here.
   quiet flag. This works around limitations with custom test harnesses
   such as cucumber.
 - Update RON to 0.7.1.
+- Improved ergonomics around `with_settings!`.  It's now a perfect match to
+  the settings object's setter methods.
+- Added `description` and `info` to snapshots. (#239)
+- Added `omit_expression` setting. (#239)
 - Added improved support for running insta from doctests. (#243)
 
 ## 1.15.0